Test overflow code into testcase
